Weekly report- updated intro, table names and footnotes as requested in GH issue 1393; added commentary for reviewers; refactored col.Monday.date lines; refactored and removed unused code in chunks 'Table6' and 'Table11'; removed unused variables created for tube_col df; clarified code for Deviation_Freq; corrected code for Fiure 6.18-6.21; created a function for figures 3.1-3.3; created functions for section6 figures; corrected figure 7.2 to include all kits not just initital kits
Weekly CSV- updated call for the polyaccession IDs, this week the variables aren't flattened with '_integer' at the end. Created a fail safe for either version of the variables

QC- updated call for the polyaccession IDs, this week the variables aren't flattened with '_integer' at the end. Created a fail safe for either version of the variables; added exclusions as per GH issues 1408, 1409, 1413; added rule updates as per GH issues 140 and 1412

SF csv- added new locations